fix вычислений с BigDecimal

+++ src/main/java/com/dorogo/MoneyCalculatorService/controllers/form/MemberForm.java
@@ -11,6 +11,6 @@ public class MemberForm {
     private String name;
 
     @Min(0L)
-    private double spent;
+    private int spent;
 
 }
